+++ src/Informedica.GenOrder.Lib/Order.fs
@@ -1,13 +1,22 @@
 ﻿namespace Informedica.GenOrder.Lib
 
+/// Types and functions that deal with an order.
+/// An `Order` models the `Prescription` of an 
+/// `Orderable` with a `StartStop` start date and
+/// stop date.
 [<CompilationRepresentation(CompilationRepresentationFlags.ModuleSuffix)>]
 module Order =
 
+    /// Contains literals used
+    /// to generate `Variable` names
     module Literals =
 
         [<Literal>]
         let adjust = "Adjust"
 
+    /// Utitlity functions to
+    /// enable mapping of a `Variable`s
+    /// to an `Order`
     module Mapping =
 
         module UN = Unit
@@ -80,10 +89,15 @@ module Order =
             | OrderableDoseAdjustRateAdjust -> "Orderable.DoseAdjust.RateAdjust"
             | AdjustQty -> "Adjust.Qty"
 
+    /// Types and functions that
+    /// model a start and stop date time
+    /// of an `Order`
     module StartStop =
         
         open System
 
+        /// There is always a `Start` or
+        /// both a `StartStop`
         type StartStop =
             | Start of DateTime
             | StartStop of DateTime * DateTime
@@ -107,75 +121,112 @@ module Order =
     module FR = VU.Frequency
     module TM = VU.Time
 
+    /// Models an order
     type Order = 
         {
+            /// Used to adjust doses
             Adjust: QT.Quantity
+            /// That what can be ordered
             Orderable: OR.Orderable
+            /// How the orderable is prescribed
             Prescription: PR.Prescription
+            /// The route of administration of the order
             Route: string // Route.T
+            /// The start stop date of the order
             StartStop: DT.StartStop
         }
 
 
-    let apply f (o: Order) = o |> f
+    /// Apply `f` to `Order` `ord`
+    let apply f (ord: Order) = ord |> f
 
+    /// Utilty function to facilitate type inference
     let get = apply id
 
-    let create adj ord pre rte sts =
+    /// Create an `Order` with
+    ///
+    /// * adj: by which doses are adjusted
+    /// * orb: the `Orderable`
+    /// * prs: `Prescription`, how the orderable is prescribed
+    /// * rot: the route of administration of the orderable
+    let create adj_qty orb prs rot sts =
         {
-            Adjust = adj 
-            Orderable = ord
-            Prescription = pre
-            Route = rte
+            Adjust = adj_qty 
+            Orderable = orb
+            Prescription = prs
+            Route = rot
             StartStop = sts
         }
 
-    let createNew adju ord prs rte = 
-        let nm = ord |> OR.getName
-        let prs = [nm] |> prs
-        let adj = adju |> QT.quantity [nm; LT.adjust]
+    /// Create a new orderable using:
+    ///
+    /// * adj_unt: the unit group by which doses are adjusted
+    /// * orb: the orderable being ordered
+    /// * str_prs: a function that takes in a list of strings 
+    /// that will generate the names and returns a `Prescription`
+    /// * rot: the route of administration
+    let createNew adj_ung orb str_prs rot = 
+        let nm = orb |> OR.getName
+        let prs = [nm] |> str_prs
+        let adj = adj_ung |> QT.quantity [nm; LT.adjust]
         let sts = DateTime.Now  |> DT.Start
-        create adj ord prs rte sts
+        create adj orb prs rot sts
 
+    /// Create a `ProductEquation` from
+    /// a `VariableUnit` list
     let toProd xs = 
         match xs with
         | h::tail -> VU.toProdEq id (fun m -> printfn "%A" m; failwith "Oops") h tail |> EQ.nonZeroOrNegative
         | _ -> failwith "Not a valid equation"
 
+    /// Create a `SumEquation` from
+    /// a `VariableUnit` list
     let toSum xs =
         match xs with
         | h::tail -> VU.toSumEq  id (fun _ -> failwith "Oops") h tail |> EQ.nonZeroOrNegative
         | _ -> failwith "Not a valid equation"
 
-    let toEqs (o: Order) =
-        let ord = o.Orderable
-        let adj = o.Adjust |> QT.toVar
-        let frq, tme = o.Prescription |> PR.toEqs
+    /// Map an `Order` *ord* to 
+    /// a list of `VariableUnit` lists
+    let toEqs (ord: Order) =
+        let orb = ord.Orderable
+        let adj = ord.Adjust |> QT.toVarUnt
+        let frq, tme = ord.Prescription |> PR.toEqs
         let hasRate = 
-            o.Prescription |> PR.isContinuous || 
-            o.Prescription |> PR.isTimed
+            ord.Prescription |> PR.isContinuous || 
+            ord.Prescription |> PR.isTimed
         
-        ord |> OR.toEqs hasRate adj frq tme
+        orb |> OR.toEqs hasRate adj frq tme
 
-    let fromEqs o vus =        
-        let ord = o.Orderable |> OR.fromEqs vus
-        let adj = o.Adjust |> QT.fromVar vus
-        let prs = o.Prescription |> PR.fromEqs vus
+    /// Map a list of `VariableUnit` lists
+    /// to an `Order` *ord*
+    let fromEqs ord vus =        
+        let orb = ord.Orderable |> OR.fromEqs vus
+        let adj = ord.Adjust |> QT.fromVar vus
+        let prs = ord.Prescription |> PR.fromEqs vus
 
         {
-            o with
+            ord with
                 Adjust = adj
-                Orderable = ord
+                Orderable = orb
                 Prescription = prs
         }
 
+    /// Turn an order into a list of string
+    /// representing variable name, valuerange 
+    /// and unit group
     let toString (ord: Order) =
         [ LT.adjust; ord.Adjust |> QT.toString ]
         |> List.append (OR.Literals.orderable::(ord.Orderable |> OR.toString))
         |> List.append ("Prescription"::(ord.Prescription |> PR.toString))
         |> List.append ("Route"::[(ord.Route)])
         
-
+    /// Solve an `Order` *ord* with
+    /// 
+    /// * n: the name of the variable to be set
+    /// * m: the mapping for the field of the order
+    /// * p: the property of the variable to be set
+    /// * v: the values to be set
     let solve n m p v u o =
         let toEql (prod, sum) =
             prod 
@@ -187,7 +238,7 @@ module Order =
 
         let toUnit n u (p, s) = 
             let us = u |> String.split "/"
-            match s::p |> VU.findVarUnit n with
+            match s::p |> VU.tryFindVarUnt n with
             | Some vu -> 
                 vu
                 |> VU.getUnitGroup