+++ RELEASE_NOTES.md
@@ -1,3 +1,9 @@
+#### 0.0.3-beta - May 25 2016
+* Updated order generation with only unit groups
+* Updated calculation model
+* Refactored code
+* Documented all code
+
 #### 0.0.2-beta - May 21 2016
 * Updated eqs (calulation model) generation
 * Improved solve syntax